requirements.txt
Flask~=3.0.0
Flask-SQLAlchemy~=3.1.1
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
pylint>=2.0.0
mailjet-rest~=1.3.4